toolkit/pyproject.toml
[tool.poetry]
name = "pythonit-toolkit"
version = "0.1.7"
description = ""
authors = ["Python Italia"]

[tool.poetry.dependencies]
python = "^3.9"
PyJWT = "^2.0.1"
starlette = {version = "^0.14.2", optional = true}

[tool.poetry.dev-dependencies]
ward = "^0.50.0-beta.0"
time-machine = "^2.1.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
